+++ test/db/postgresql/matcher_test.dart
@@ -38,6 +38,11 @@ void main() {
       var results = await q.fetch();
       expect(results.length, 1);
       expect(results.first.id, 1);
+
+      q = new Query<TestModel>()..where["id"] = whereNot(whereEqualTo(1));
+      results = await q.fetch();
+      expect(results.length, 5);
+      expect(results.any((tm) => tm.id == 1), false);
     });
 
     test("String value, case sensitive default", () async {
@@ -46,9 +51,18 @@ void main() {
       expect(results.length, 1);
       expect(results.first.id, 1);
 
+      q = new Query<TestModel>()..where["email"] = whereNot(whereEqualTo("0@a.com"));
+      results = await q.fetch();
+      expect(results.length, 5);
+      expect(results.any((tm) => tm.id == 1), false);
+
       q = new Query<TestModel>()..where["email"] = whereEqualTo("0@A.com");
       results = await q.fetch();
       expect(results.length, 0);
+
+      q = new Query<TestModel>()..where["email"] = whereNot(whereEqualTo("0@A.com"));
+      results = await q.fetch();
+      expect(results.length, 6);
     });
 
     test("String value, case sensitive default", () async {
@@ -56,6 +70,11 @@ void main() {
       var results = await q.fetch();
       expect(results.length, 1);
       expect(results.first.id, 1);
+
+      q = new Query<TestModel>()..where["email"] = whereNot(whereEqualTo("0@A.com", caseSensitive: false));
+      results = await q.fetch();
+      expect(results.length, 5);
+      expect(results.any((tm) => tm.email == "0@a.com"), false);
     });
   });
 
@@ -66,6 +85,11 @@ void main() {
     expect(results.length, 2);
     expect(results.first.id, 1);
     expect(results.last.id, 2);
+
+    q = new Query<TestModel>()..where["id"] = whereNot(whereLessThan(3));
+    results = await q.fetch();
+    expect(results.length, 4);
+    expect(results.every((tm) => tm.id >= 3), true);
   });
 
   test("Less than equal to matcher", () async {
@@ -75,6 +99,11 @@ void main() {
     expect(results[0].id, 1);
     expect(results[1].id, 2);
     expect(results[2].id, 3);
+
+    q = new Query<TestModel>()..where["id"] = whereNot(whereLessThanEqualTo(3));
+    results = await q.fetch();
+    expect(results.length, 3);
+    expect(results.every((tm) => tm.id > 3), true);
   });
 
   test("Greater than matcher", () async {
@@ -83,6 +112,11 @@ void main() {
     expect(results.length, 2);
     expect(results[0].id, 5);
     expect(results[1].id, 6);
+
+    q = new Query<TestModel>()..where["id"] = whereNot(whereGreaterThan(4));
+    results = await q.fetch();
+    expect(results.length, 4);
+    expect(results.every((tm) => tm.id <= 4), true);
   });
 
   test("Greater than equal to matcher", () async {
@@ -92,6 +126,11 @@ void main() {
     expect(results[0].id, 4);
     expect(results[1].id, 5);
     expect(results[2].id, 6);
+
+    q = new Query<TestModel>()..where["id"] = whereNot(whereGreaterThanEqualTo(4));
+    results = await q.fetch();
+    expect(results.length, 3);
+    expect(results.every((tm) => tm.id < 4), true);
   });
 
   group("Not equal matcher", () {
@@ -100,6 +139,11 @@ void main() {
       var results = await q.fetch();
       expect(results.length, 5);
       expect(results.any((t) => t.id == 1), false);
+
+      q = new Query<TestModel>()..where["id"] = whereNot(whereNotEqualTo(1));
+      results = await q.fetch();
+      expect(results.length, 1);
+      expect(results.any((t) => t.id == 1), true);
     });
 
     test("String value, case sensitive default", () async {
@@ -108,9 +152,18 @@ void main() {
       expect(results.length, 5);
       expect(results.any((t) => t.id == 1), false);
 
+      q = new Query<TestModel>()..where["email"] = whereNot(whereNotEqualTo("0@a.com"));
+      results = await q.fetch();
+      expect(results.length, 1);
+      expect(results.any((t) => t.id == 1), true);
+
       q = new Query<TestModel>()..where["email"] = whereNotEqualTo("0@A.com");
       results = await q.fetch();
       expect(results.length, 6);
+
+      q = new Query<TestModel>()..where["email"] = whereNot(whereNotEqualTo("0@A.com"));
+      results = await q.fetch();
+      expect(results.length, 0);
     });
 
     test("String value, case sensitive default", () async {
@@ -118,6 +171,11 @@ void main() {
       var results = await q.fetch();
       expect(results.length, 5);
       expect(results.any((t) => t.id == 1), false);
+
+      q = new Query<TestModel>()..where["email"] = whereNot(whereNotEqualTo("0@A.com", caseSensitive: false));
+      results = await q.fetch();
+      expect(results.length, 1);
+      expect(results.any((t) => t.id == 1), true);
     });
   });
 
@@ -143,6 +201,15 @@ void main() {
     expect(results[0].id, 2);
     expect(results[1].id, 3);
     expect(results[2].id, 4);
+
+    q = new Query<TestModel>()..where["id"] = whereNot(whereBetween(2, 4));
+    results = await q.fetch();
+    expect(results.length, 3);
+
+    results.sort((t1, t2) => t1.id.compareTo(t2.id));
+    expect(results[0].id, 1);
+    expect(results[1].id, 5);
+    expect(results[2].id, 6);
   });
 
   test("whereOutsideOf matcher", () async {
@@ -152,6 +219,15 @@ void main() {
     expect(results[0].id, 1);
     expect(results[1].id, 5);
     expect(results[2].id, 6);
+
+    q = new Query<TestModel>()..where["id"] = whereNot(whereOutsideOf(2, 4));
+    results = await q.fetch();
+    expect(results.length, 3);
+
+    results.sort((t1, t2) => t1.id.compareTo(t2.id));
+    expect(results[0].id, 2);
+    expect(results[1].id, 3);
+    expect(results[2].id, 4);
   });
 
   test("whereRelatedByValue matcher", () async {
@@ -159,6 +235,11 @@ void main() {
     var results = await q.fetch();
     expect(results.length, 1);
     expect(results.first.owner.id, 1);
+
+    // Does not include null values; this is intentional.
+    q = new Query<InnerModel>()..where["owner"] = whereNot(whereRelatedByValue(1));
+    results = await q.fetch();
+    expect(results.length, 0);
   });
 
   test("whereNull matcher", () async {
@@ -166,6 +247,11 @@ void main() {
     var results = await q.fetch();
     expect(results.length, 1);
     expect(results.first.name, "No one's");
+
+    q = new Query<InnerModel>()..where["owner"] = whereNot(whereNull);
+    results = await q.fetch();
+    expect(results.length, 1);
+    expect(results.first.name, "Bob's");
   });
 
   test("whereNotNull matcher", () async {
@@ -173,19 +259,11 @@ void main() {
     var results = await q.fetch();
     expect(results.length, 1);
     expect(results.first.name, "Bob's");
-  });
 
-  test("whereAnyMatch matcher", () async {
-    var q = new Query<TestModel>()..join(object: (t) => t.inner);
-    var results = await q.fetch();
-    expect(results.length, 6);
-
-    expect(results.first.name, "Bob");
-    expect(results.first.inner.name, "Bob's");
-
-    for (var i = 1; i < results.length; i++) {
-      expect(results[i].inner, isNull);
-    }
+    q = new Query<InnerModel>()..where["owner"] = whereNot(whereNotNull);
+    results = await q.fetch();
+    expect(results.length, 1);
+    expect(results.first.name, "No one's");
   });
 
   group("whereContains matcher", () {
@@ -195,6 +273,12 @@ void main() {
       expect(results.length, 2);
       expect(results.first.name, "Sally");
       expect(results.last.name, "Kanye");
+
+      q = new Query<TestModel>()..where["name"] = whereNot(whereContainsString("y"));
+      results = await q.fetch();
+      expect(results.length, 4);
+      expect(results.any((tm) => tm.name == "Sally"), false);
+      expect(results.any((tm) => tm.name == "Kanye"), false);
     });
 
     test("Case insensitive", () async {
@@ -203,6 +287,12 @@ void main() {
       expect(results.length, 2);
       expect(results.first.name, "Sally");
       expect(results.last.name, "Kanye");
+
+      q = new Query<TestModel>()..where["name"] = whereNot(whereContainsString("Y", caseSensitive: false));
+      results = await q.fetch();
+      expect(results.length, 4);
+      expect(results.any((tm) => tm.name == "Sally"), false);
+      expect(results.any((tm) => tm.name == "Kanye"), false);
     });
   });
 
@@ -212,6 +302,11 @@ void main() {
       var results = await q.fetch();
       expect(results.length, 1);
       expect(results.first.name, "Bob");
+
+      q = new Query<TestModel>()..where["name"] = whereNot(whereBeginsWith("B"));
+      results = await q.fetch();
+      expect(results.length, 5);
+      expect(results.any((tm) => tm.name == "Bob"), false);
     });
 
     test("Case insensitive", () async {
@@ -219,6 +314,11 @@ void main() {
       var results = await q.fetch();
       expect(results.length, 1);
       expect(results.first.name, "Bob");
+
+      q = new Query<TestModel>()..where["name"] = whereNot(whereBeginsWith("b", caseSensitive: false));
+      results = await q.fetch();
+      expect(results.length, 5);
+      expect(results.any((tm) => tm.name == "Bob"), false);
     });
   });
 
@@ -228,6 +328,11 @@ void main() {
       var results = await q.fetch();
       expect(results.length, 1);
       expect(results.first.name, "Tim");
+
+      q = new Query<TestModel>()..where["name"] = whereNot(whereEndsWith("m"));
+      results = await q.fetch();
+      expect(results.length, 5);
+      expect(results.any((tm) => tm.name == "Tim"), false);
     });
 
     test("Case insensitive", () async {
@@ -235,6 +340,11 @@ void main() {
       var results = await q.fetch();
       expect(results.length, 1);
       expect(results.first.name, "Tim");
+
+      q = new Query<TestModel>()..where["name"] = whereNot(whereEndsWith("M", caseSensitive: false));
+      results = await q.fetch();
+      expect(results.length, 5);
+      expect(results.any((tm) => tm.name == "Tim"), false);
     });
   });
 
@@ -247,6 +357,12 @@ void main() {
       expect(results.length, 4);
       expect(results.any((t) => t.name == "Sally"), false);
       expect(results.any((t) => t.name == "Kanye"), false);
+
+      q = new Query<TestModel>()..where["name"] = whereNot(whereDoesNotContain("y"));
+      results = await q.fetch();
+      expect(results.length, 2);
+      expect(results.any((t) => t.name == "Sally"), true);
+      expect(results.any((t) => t.name == "Kanye"), true);
     });
 
     test("Case insensitive", () async {
@@ -255,6 +371,12 @@ void main() {
       expect(results.length, 4);
       expect(results.any((t) => t.name == "Sally"), false);
       expect(results.any((t) => t.name == "Kanye"), false);
+
+      q = new Query<TestModel>()..where["name"] = whereNot(whereDoesNotContain("Y", caseSensitive: false));
+      results = await q.fetch();
+      expect(results.length, 2);
+      expect(results.any((t) => t.name == "Sally"), true);
+      expect(results.any((t) => t.name == "Kanye"), true);
     });
   });
 
@@ -264,6 +386,11 @@ void main() {
       var results = await q.fetch();
       expect(results.length, 5);
       expect(results.any((t) => t.name == "Bob"), false);
+
+      q = new Query<TestModel>()..where["name"] = whereNot(whereDoesNotBeginWith("B"));
+      results = await q.fetch();
+      expect(results.length, 1);
+      expect(results.any((t) => t.name == "Bob"), true);
     });
 
     test("Case insensitive", () async {
@@ -271,6 +398,11 @@ void main() {
       var results = await q.fetch();
       expect(results.length, 5);
       expect(results.any((t) => t.name == "Bob"), false);
+
+      q = new Query<TestModel>()..where["name"] = whereNot(whereDoesNotBeginWith("b", caseSensitive: false));
+      results = await q.fetch();
+      expect(results.length, 1);
+      expect(results.any((t) => t.name == "Bob"), true);
     });
   });
 
@@ -280,6 +412,11 @@ void main() {
       var results = await q.fetch();
       expect(results.length, 5);
       expect(results.any((t) => t.name == "Tim"), false);
+
+      q = new Query<TestModel>()..where["name"] = whereNot(whereDoesNotEndWith("m"));
+      results = await q.fetch();
+      expect(results.length, 1);
+      expect(results.any((t) => t.name == "Tim"), true);
     });
 
     test("Case insensitive", () async {
@@ -287,6 +424,11 @@ void main() {
       var results = await q.fetch();
       expect(results.length, 5);
       expect(results.any((t) => t.name == "Tim"), false);
+
+      q = new Query<TestModel>()..where["name"] = whereNot(whereDoesNotEndWith("M", caseSensitive: false));
+      results = await q.fetch();
+      expect(results.length, 1);
+      expect(results.any((t) => t.name == "Tim"), true);
     });
   });